refactor(types): cambiar tipos de precios y cantidades de double a BigDecimal/string
Cambia los tipos de datos para precios y cantidades en los archivos proto de double a string, y en los DTOs y servicios de Java de Double a BigDecimal. Esto mejora la precisión en los cálculos decimales y evita problemas de redondeo asociados con tipos flotantes. Los mappers se actualizan para manejar las conversiones apropiadas entre string y BigDecimal.

+++ servicio-catalogo/src/main/java/com/isam/dto/oferta/OfertaDto.java
@@ -14,7 +14,7 @@ public record OfertaDto(
     @NotNull(message = "El precio promocional es obligatorio")
     @DecimalMin(value = "0.0", inclusive = false, message = "El precio promocional debe ser mayor que 0")
     @Digits(integer = 10, fraction = 2, message = "El precio debe tener máximo 10 dígitos enteros y 2 decimales")
-    Double precioPromocional,
+    BigDecimal precioPromocional,
     
     @NotBlank(message = "El tipo de promoción es obligatorio")
     @Size(max = 50, message = "El tipo de promoción no puede exceder 50 caracteres")
