+++ servicio-catalogo/src/main/java/com/isam/dto/producto/ProductoDto.java
@@ -1,5 +1,6 @@
 package com.isam.dto.producto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.isam.dto.categoria.CategoriaDto;
@@ -10,19 +11,19 @@ import jakarta.validation.constraints.NotNull;
 public record ProductoDto(
         @NotNull(message = "El SKU es obligatorio")
         String sku,
-        
+
         String ean,
-        
+
         String plu,
-        
+
         @NotNull(message = "El nombre es obligatorio")
         String nombre,
-        
+
         String descripcion,
-        
+
         @NotNull(message = "El precio de venta es obligatorio")
-        Double precioVenta,
-        
+        BigDecimal precioVenta,
+
         @NotNull(message = "El campo 'caduca' es obligatorio")
         Boolean caduca,
