+++ servicio-inventario/src/main/java/com/isam/dto/inventario/DetallesInventarioCompletoDto.java
@@ -1,5 +1,6 @@
 package com.isam.dto.inventario;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.isam.dto.lote.DetalleLoteDto;
@@ -8,8 +9,8 @@ import com.isam.model.UnidadMedida;
 public record DetallesInventarioCompletoDto(
     String sku,
     String nombreProducto,
-    Double stockTotalAlmacen,
-    Double stockTotalEstanteria,
+    BigDecimal stockTotalAlmacen,
+    BigDecimal stockTotalEstanteria,
     UnidadMedida unidadMedida,
     List<DetalleLoteDto> lotes
 ){}
