+++ servicio-inventario/src/main/java/com/isam/dto/lote/LoteDto.java
@@ -3,15 +3,16 @@ package com.isam.dto.lote;
 import com.isam.model.EstadoLote;
 import com.isam.model.UnidadMedida;
 
+import java.math.BigDecimal;
 
 public record LoteDto(
     String idLote,
     String sku,
     String idInventario,
     String numeroLote,
-    Double cantidadEntrada,
-    Double cantidadAlmacen,
-    Double cantidadEstanteria,
+    BigDecimal cantidadEntrada,
+    BigDecimal cantidadAlmacen,
+    BigDecimal cantidadEstanteria,
     String fechaCaducidad,  // YYYY-MM-DD
     String idProveedor,
     String fechaIngreso,    // YYYY-MM-DD
